+++ CHANGELOG.md
@@ -7,3 +7,12 @@
 ### Added
 
 -   MINOR: new `hasFlowConfig()` function
+
+
+### Changed
+
+-   MAJOR: use `async`/`await`, require Node.js 8.x and newer
+
+-   CHORE: no longer depend upon [p-filter](https://www.npmjs.com/package/p-filter)
+
+-   CHORE: no longer depend upon [pify](https://www.npmjs.com/package/pify)